+++ R/sjplot_themes.R
@@ -33,11 +33,26 @@
 #'          two-element numeric vector with values from 0-1.
 #' @param inside Logical, use \code{TRUE} to put legend inside the plotting area.
 #'        See also \code{pos}.
-#' @param theme Optional ggplot-theme-object, which is needed in case multiple
+#' @param base.theme Optional ggplot-theme-object, which is needed in case multiple
 #'        functions should be combined, e.g. \code{theme_sjplot() + label_angle()}.
-#'        In such cases, use \code{label_angle(theme = theme_sjplot())}.
-
-
+#'        In such cases, use \code{label_angle(base.theme = theme_sjplot())}.
+#'
+#' @examples
+#' # prepare data
+#' library(sjmisc)
+#' data(efc)
+#' efc <- to_factor(efc, c161sex, e42dep, c172code)
+#' m <- lm(neg_c_7 ~ pos_v_4 + c12hour + e42dep + c172code, data = efc)
+#'
+#' # create plot-object
+#' p <- plot_model(m)
+#'
+#' # change theme
+#' p + theme_sjplot()
+#'
+#' # change font-size
+#' p + font_size(axis_title.x = 30)
+#'
 #' @rdname sjPlot-themes
 #' @export
 theme_sjplot <- function(base_size = 12, base_family = "") {
@@ -107,10 +122,10 @@ theme_538 <- function(base_size = 12, base_family = "") {
 
 #' @rdname sjPlot-themes
 #' @export
-font_size <- function(title, axis_title.x, axis_title.y, labels.x, labels.y, offset.x, offset.y, theme) {
+font_size <- function(title, axis_title.x, axis_title.y, labels.x, labels.y, offset.x, offset.y, base.theme) {
   # get current theme
-  if (!missing(theme))
-    cur.theme <- theme
+  if (!missing(base.theme))
+    cur.theme <- base.theme
   else
     cur.theme <- theme_get()
 
@@ -155,10 +170,10 @@ font_size <- function(title, axis_title.x, axis_title.y, labels.x, labels.y, off
 
 #' @rdname sjPlot-themes
 #' @export
-label_angle <- function(angle.x, angle.y, theme) {
+label_angle <- function(angle.x, angle.y, base.theme) {
   # get current theme
-  if (!missing(theme))
-    cur.theme <- theme
+  if (!missing(base.theme))
+    cur.theme <- base.theme
   else
     cur.theme <- theme_get()
 
@@ -178,10 +193,10 @@ label_angle <- function(angle.x, angle.y, theme) {
 
 #' @rdname sjPlot-themes
 #' @export
-legend_style <- function(inside, pos, justify, theme) {
+legend_style <- function(inside, pos, justify, base.theme) {
   # get current theme
-  if (!missing(theme))
-    cur.theme <- theme
+  if (!missing(base.theme))
+    cur.theme <- base.theme
   else
     cur.theme <- theme_get()